+++ App/Resources/tarea_resource.py
@@ -2,10 +2,15 @@ from Models.tarea_model import Tarea
 from flask import jsonify, request
 from Schemas.tarea_schema import Tarea_schema, Tareas_schema
 from config import db
-# from flask_jwt_extended import create_access_token, jwt_required - Pendiente
+from flask_jwt_extended import get_jwt, jwt_required
 
+@jwt_required()
 def get_tarea(id):
     try:
+        claims = get_jwt()
+        if claims.get('role') not in ['normal', 'admin']:
+            return jsonify({"message": "Role required"}), 403
+        
         tarea = Tarea.query.get(id)
         if tarea:
             return jsonify(Tarea_schema.dump(tarea))
@@ -15,15 +20,26 @@ def get_tarea(id):
     except Exception as e: 
         return jsonify({"message": f"Error: {str(e)}"}), 500
     
+@jwt_required()
 def get_tareas():
     try: 
+
+        claims = get_jwt()
+        if claims.get('role') not in ['normal', 'admin']:
+            return jsonify({"message": "Role required"}), 403
+        
         tareas = Tarea.query.all()
         return jsonify(Tareas_schema.dump(tareas))
     except Exception as e:
         return jsonify({"message": f"Error: {str(e)}"}), 500
-    
+
+@jwt_required()
 def create_tarea():
     try: 
+        claims = get_jwt()
+        if claims.get('role') not in ['normal', 'admin']:
+            return jsonify({"message": "Role required"}), 403
+        
         data = request.get_json()
         if not data:
             return jsonify({"message": "Error: No data received"}), 400
@@ -38,9 +54,15 @@ def create_tarea():
     except Exception as e:
         db.session.rollback()
         return jsonify({"message": f"Error: {str(e)}"}), 500
-    
+
+@jwt_required()
 def update_tarea(id):
     try:
+
+        claims = get_jwt()
+        if claims.get('role') not in ['normal', 'admin']:
+            return jsonify({"message": "Role required"}), 403
+        
         tarea = Tarea.query.get(id)
         data = request.get_json()
         if not data: 
@@ -56,9 +78,14 @@ def update_tarea(id):
     except Exception as e:
         db.session.rollback()
         return jsonify({"message": f"Error: {str(e)}"}), 500
-    
+
+@jwt_required()
 def delete_tarea(id):
     try:
+        claims = get_jwt()
+        if claims.get('role') not in ['normal', 'admin']:
+            return jsonify({"message": "Role required"}), 403
+        
         tarea = Tarea.query.get(id)
         db.session.delete(tarea)
         db.session.commit()
